Remove platform specific app implementations
